add JWT Token auth

+++ src/controller/place-controller.ts
@@ -15,7 +15,7 @@ const getAllPlaces = (req: express.Request, res: express.Response, next: express
             console.log('err', err);
             return next(new HttpError('Could not find places', 500));
         }
-        res.json({ status:'success',data:{places: places} });
+        res.json({ status: 'success', data: { places: places } });
     })
 };
 
@@ -26,7 +26,7 @@ const getPlaceById = (req: express.Request, res: express.Response, next: express
         if (err) {
             return next(new HttpError('Could not find place', 500));
         }
-        res.json({ status:'success',data:{place: place} });
+        res.json({ status: 'success', data: { place: place } });
     })
 };
 
@@ -37,7 +37,7 @@ const getPlacesByUserId = (req: express.Request, res: express.Response, next: ex
         if (err) {
             return next(new HttpError('Could not find places', 500));
         }
-        res.json({status:'success',data:{ places: places} });
+        res.json({ status: 'success', data: { places: places } });
     })
 }
 const createNewPlace = async (req: express.Request, res: express.Response, next: express.NextFunction) => {
@@ -49,13 +49,15 @@ const createNewPlace = async (req: express.Request, res: express.Response, next:
         return next(error);
     }
     const body = req.body as Place;
-    console.log('body',body);
+    console.log('body', body);
     const { title, description, location, address, creator, image } = body;
     const id = crypto.randomBytes(16).toString("hex");
     console.log(body)
     let coordinates;
     try {
-        coordinates = await getCoordsForAddress(address);
+        //in case we want to use google api
+        // coordinates = await getCoordsForAddress(address);
+        coordinates = { lat: 40.7484474, lng: -73.9871516 };
     } catch (error) {
         console.log(error)
         return next(error);
@@ -72,12 +74,12 @@ const createNewPlace = async (req: express.Request, res: express.Response, next:
     let user;
     try {
         user = await UserModel.findById(creator);
-    }catch(err){
+    } catch (err) {
         let error = new HttpError('error in create place', 500);
         return next(error);
     }
     console.log(user);
-    if(!user){
+    if (!user) {
         let error = new HttpError('Could not find user', 500);
         return next(error);
     }
@@ -85,7 +87,7 @@ const createNewPlace = async (req: express.Request, res: express.Response, next:
     //1 create  place and check if it is created or not
     try {
         await place.save();
-    }catch(err){
+    } catch (err) {
         let error = new HttpError('error in create place', 500);
         console.log('err', err);
         return next(error);
@@ -94,15 +96,15 @@ const createNewPlace = async (req: express.Request, res: express.Response, next:
     user?.places?.push(place);
     //3 save user
     await user.save()
-    res.status(201).json({ 
-        status:'success',
-        data:{
-            place: createdPlace 
+    res.status(201).json({
+        status: 'success',
+        data: {
+            place: createdPlace
         }
     });
 }
 
-const updatePlaceById = (req: express.Request, res: express.Response, next: express.NextFunction) => {
+const updatePlaceById = async (req: express.Request, res: express.Response, next: express.NextFunction) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         console.log(errors);
@@ -113,44 +115,64 @@ const updatePlaceById = (req: express.Request, res: express.Response, next: expr
     const body = req.body as Place;
     const { title, description } = body;
     console.log(body);
-    PlaceModel
-        .updateOne({ _id: placeId }, { title, description })
-        .then((result) => {
-            res.status(200).json({ status:'success',data:{
+    let place;
+    try {
+        place = await PlaceModel.findById(placeId)
+    } catch (err) {
+        let error = new HttpError('Could not find place', 500);
+        return next(error);
+    }
+    console.log('userData', req.body.userData)
+    if (place?.creator?.toString() != req.body.userData.userId) {
+        let error = new HttpError('You are not Allowed to edit this place', 401);
+        return next(error);
+    }
+
+    try {
+        const result = await PlaceModel.updateOne({ _id: placeId }, { title, description })
+        res.status(200).json({
+            status: 'success', data: {
                 result
-            } });
-        })
-        .catch((err) => {
-            let error = new HttpError('Could not update place', 500);
-            return next(error);
-        })
+            }
+        });
+    } catch (err) {
+        let error = new HttpError('Could not update place', 500);
+        return next(error);
+    }
+
 }
 
-const deletePlaceById = async(req: express.Request, res: express.Response, next: express.NextFunction) => {
-    const params  = req.params as RequestParams;
+const deletePlaceById = async (req: express.Request, res: express.Response, next: express.NextFunction) => {
+    const params = req.params as RequestParams;
     const placeId = params.placeId;
     let place;
     console.log(placeId);
     place = await PlaceModel.findById(placeId)
-    if(!place){
+    if (!place) {
         let error = new HttpError('Could not find place', 500);
         return next(error);
     }
-    const imagePath:any = place.image;
+
+    if (place?.creator?.toString() != req.body.userData.userId) {
+        let error = new HttpError('You are not Allowed to delete this place', 401);
+        return next(error);
+    }
+
+    const imagePath: any = place.image;
     // let place;
     try {
         // place = await PlaceModel.findById(placeId).populate('creator');
         await PlaceModel.deleteOne({ _id: placeId });
-    }catch(err){
+    } catch (err) {
         let error = new HttpError('Could not delete place', 500);
         return next(error);
-    }    
+    }
     // await UserModel.updateMany({}, { $pull: { 'users.places': placeId } })
     await UserModel.updateMany({}, { $pull: { places: placeId } });
-   fs.unlink(imagePath,(err:any)=>{
-         console.log(err);
-   });
-    res.status(200).json({status:'success', message: 'Deleted place' });
+    fs.unlink(imagePath, (err: any) => {
+        console.log(err);
+    });
+    res.status(200).json({ status: 'success', message: 'Deleted place' });
 }
 
 export { getPlaceById, getPlacesByUserId, createNewPlace, deletePlaceById, updatePlaceById, getAllPlaces };
